Add MSVC build to CI (#159)

+++ frontend/mp4read.h
@@ -41,7 +41,7 @@ typedef struct
     // sample depth
     uint32_t bits;
     // buffer config
-    uint16_t buffersize;
+    uint32_t buffersize;
     uint32_t bitratemax;
     uint32_t bitrateavg;
     // frame size / offsets
@@ -51,17 +51,17 @@ typedef struct
         slice_info_t *map;
         uint32_t nsamples;
         uint32_t nsclices;
-        int current;
-        int maxsize;
+        uint32_t current;
+        uint32_t maxsize;
     } frame;
     // AudioSpecificConfig data:
     struct
     {
         uint8_t buf[10];
-        int size;
+        uint32_t size;
     } asc;
     struct {
-        int size;
+        uint32_t size;
         uint8_t *data;
     } bitbuf;
     struct {
@@ -73,6 +73,6 @@ typedef struct
 extern mp4config_t mp4config;
 
 int mp4read_open(char *name);
-int mp4read_seek(int framenum);
+int mp4read_seek(uint32_t framenum);
 int mp4read_frame(void);
 int mp4read_close(void);
